Python/APP1_python_bootcamp_1/Python_Bootcamp.Day_06-2/src/requirements.txt
grpcio==1.69.0
protobuf==5.29.3
pydantic==2.10.5
SQLAlchemy==2.0.37
psycopg2
grpcio-tools